Fixed bug in pledges/ not displaying, updated missing field for categories
Bug fixed for pledges. For some reason it didn't like PledgeDetailSerializer, so once this was deleted the 500server error was fixed and I could load the host server again now issues. Also implemented fix to categories field so it actually displays on my projects now. Still trying to get a particular update field for that.